% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsflabel.R
\docType{package}
\name{ggsflabel}
\alias{ggsflabel}
\alias{ggsflabel-package}
\title{Labels for 'sf' with 'ggplot2'}
\description{
Provides several 'geom's which plot labels generated from 'sf' data.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com}

}
