% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims.R
\name{lims_bbox}
\alias{lims_bbox}
\title{Limits for 'sf'}
\usage{
lims_bbox(x)
}
\arguments{
\item{x}{An object of \code{sf}.}
}
\description{
Set scale limits based on the 'bbox' of 'sf' object.
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
points_sfg <- sf::st_multipoint(as.matrix(expand.grid(x = -90:-70, y = 30:40)))
points_sfc <- sf::st_sfc(points_sfg, crs = sf::st_crs(nc))

p <- ggplot() +
  geom_sf(data = nc, aes(fill = AREA)) +
  geom_sf(data = points_sfc)

# too wide
p

# shrink the limits to the bbox of nc
p + lims_bbox(nc)

}
